Struct gw2_mumble::MumblePtr

pub struct MumblePtr(/* private fields */);

A pointer to LinkedMem with utility.

Implementations

impl MumblePtr

pub unsafe fn new(ptr: *mut LinkedMem) -> Option<Self>

Creates a new access point to the LinkedMem.

Safety

If the passed pointer is non-null, it must be properly aligned, dereferenceable, and point to an initialized instance of LinkedMem.

pub const fn as_non_null(&self) -> NonNull<LinkedMem>

Returns a NonNull pointer to the LinkedMem.

pub const fn as_ptr(&self) -> *const LinkedMem

Returns a raw pointer to the LinkedMem.

pub const unsafe fn as_ref(&self) -> &LinkedMem

Returns a reference to the LinkedMem.

Safety

See NonNull::as_ref.

pub fn read(&self) -> LinkedMem

Reads the entire current LinkedMem contents.

pub fn read_ui_version(&self) -> u32

Reads the current ui_version.

pub fn read_ui_tick(&self) -> u32

Reads the current ui_tick.

pub fn read_avatar(&self) -> Position

Reads the current player avatar Position.

pub fn read_name(&self) -> Vec<u16>

Reads the current game name.

pub fn read_name_string(&self) -> OsString

Reads the current game name as OsString.

pub fn read_camera(&self) -> Position

Reads the current camera Position.

pub fn read_identity(&self) -> Vec<u16>

Reads the current player identity.

pub fn read_identity_string(&self) -> OsString

Reads the current player identity as OsString.

pub fn parse_identity(&self) -> Result<Identity>

Parses the current player identity JSON contents.

pub fn read_context(&self) -> Context

Reads the current Context.

pub fn read_server_address(&self) -> [u8; 28]

Reads the current server address.

pub fn read_map_id(&self) -> u32

Reads the current map id.

pub fn read_map_type(&self) -> u32

Reads the current map type.

pub fn read_shard_id(&self) -> u32

Reads the current shard id.

pub fn read_instance(&self) -> u32

Reads the current instance id.

pub fn read_build_id(&self) -> u32

Reads the current build id.

pub fn read_ui_state(&self) -> UiState

Reads the current UI state.

pub fn read_compass_width(&self) -> u16

Reads the current compass width in pixels.

pub fn read_compass_height(&self) -> u16

Reads the current compass height in pixels.

pub fn read_compass_dimensions(&self) -> [u16; 2]

Reads the current compass dimensions in pixels.

pub fn read_compass_rotation(&self) -> f32

Reads the current compass rotation in radians.

pub fn read_player_x(&self) -> f32

Reads the current player position x in continent coordinates.

Not updated in competitive modes.

pub fn read_player_y(&self) -> f32

Reads the current player position y in continent coordinates.

Not updated in competitive modes.

pub fn read_player_position(&self) -> [f32; 2]

Reads the current player position x in continent coordinates.

Not updated in competitive modes.

pub fn read_map_center_x(&self) -> f32

Reads the current map center x in continent coordinates.

Not updated in competitive modes.

pub fn read_map_center_y(&self) -> f32

Reads the current map center y in continent coordinates.

Not updated in competitive modes.

pub fn read_map_center(&self) -> [f32; 2]

Reads the current map center in continent coordinates.

Not updated in competitive modes.

pub fn read_map_scale(&self) -> f32

Reads the map scale.

pub fn read_process_id(&self) -> u32

Reads the process id.

pub fn read_mount_index(&self) -> Mount

Reads the currently used mount.

pub fn read_description(&self) -> Vec<u16>

Reads the game description.

pub fn read_description_string(&self) -> OsString

Reads the game description as OsString.